authServer/requirements.txt
bcrypt
fastapi
uvicorn
sqlalchemy
python-jose
python-multipart
cryptography
passlib
psycopg2